requirements.txt
certifi==2020.12.5
chardet==4.0.0
docutils==0.16
future==0.18.2
gi==1.2
idna==2.10
iso8601==0.1.13
Kivy==2.0.0
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
numpy==1.19.5
opencv-python==4.5.1.48
picamera==1.13
Pillow==8.1.0
Pygments==2.7.4
pyserial==3.5
PyYAML==5.4.1
requests==2.25.1
RPi.GPIO==0.7.0
urllib3==1.26.3
